Refactor startup for cypress
